Unify the enumerators
